// src/types/student-perform.enum.ts
export enum StudentPerformance {
  A = 'A', // Отлично. Ученик хорошо понимает предмет и внимателен к деталям.
  B = 'B', // Хорошо. Работа выполнена на достойном уровне, в основном без ошибок.
  C = 'C', // Удовлетворительно. Немного не дотягивает до необходимого уровня знаний.
  D = 'D', // Неудовлетворительная работа. Много ошибок, неточностей. Ученик рискует не справиться с дисциплиной.
  U = 'U', // Чрезвычайно низкий результат. Ставится в редких случаях.
}